mlprocessing/requirements.txt
opencv-python-headless
psycopg2-binary
redis
ultralytics
